feat: add request context service

+++ src/modules/comments/comments.service.ts
@@ -1,4 +1,5 @@
 import { Injectable, NotFoundException } from '@nestjs/common'
+import { RequestContextService } from 'src/common/services/request-context/request-context.service'
 import { PrismaService } from 'src/prisma.service'
 import { AddCommentDTO, UpdateCommentDTO } from './comments.dto'
 
@@ -13,14 +14,18 @@ const authorAttributes = {
 
 @Injectable()
 export class CommentsService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly requestContext: RequestContextService,
+  ) {}
 
-  create({ authorId, data, taskId }: { data: AddCommentDTO; taskId: string; authorId: string }) {
+  create({ data, taskId }: { data: AddCommentDTO; taskId: string }) {
+    const userId = this.requestContext.getUserId()
     return this.prisma.comment.create({
       data: {
         ...data,
         task: { connect: { id: taskId } },
-        author: { connect: { id: authorId } },
+        author: { connect: { id: userId } },
       },
       include: { author: authorAttributes },
     })
@@ -57,11 +62,13 @@ export class CommentsService {
   }
 
   async update({ data, id }: { data: UpdateCommentDTO; id: string }) {
+    const userId = this.requestContext.getUserId()
     const comment = await this.prisma.comment.findFirst({ where: { id } })
     if (!comment) throw new NotFoundException('Comment not found')
     return this.prisma.comment.update({
       where: {
         id,
+        authorId: userId,
       },
       data,
       include: { author: authorAttributes },
@@ -69,9 +76,12 @@ export class CommentsService {
   }
 
   async delete(id: string) {
-    const comment = await this.prisma.comment.findFirst({ where: { id } })
+    const userId = this.requestContext.getUserId()
+    const comment = await this.prisma.comment.findFirst({ where: { id, authorId: userId } })
+
     if (!comment) throw new NotFoundException('Comment not found')
-    await this.prisma.comment.delete({ where: { id } })
+
+    await this.prisma.comment.delete({ where: { id, authorId: userId } })
     return
   }
 }
